Added default text to the homepage
Usually, the user's last or current lesson should be displayed in the
content area. However, for new or un logged in users, the default text
will greet them and show them how to use the site.

Additionally, the default text provides HTML elements with unique tags
for the javascript to target.